refactor: split edge attachment resolution + registry backend factory (#545)
- agent/internal/edge/gatewayapi/attachment: attachment/ReferenceGrant/
  hostname resolution extracted from the reconciler, proxy-free (moved
  code, behavior identical); its tests move with it
- registry/backend: single factory selecting ddb/etcd/k8s/registrar so
  the base registry package (and the agent binary) no longer links every
  backend, the AWS SDK, or controller-runtime
- registry/registrarclient: the agent's registrar-backend constructor,
  the only backend the node agent links
