+++ registry/BUILD.bazel
@@ -4,10 +4,6 @@ go_library(
     name = "registry",
     srcs = [
         "cni.go",
-        "ddb.go",
-        "etcd.go",
-        "k8s.go",
-        "registrar.go",
         "registry.go",
     ],
     importpath = "github.com/bpalermo/aether/registry",
@@ -18,12 +14,6 @@ go_library(
         "//common/constants",
         "//common/serviceref",
         "//registry/export",
-        "//registry/internal/ddb",
-        "//registry/internal/etcd",
-        "//registry/internal/k8s",
-        "//registry/internal/registrar",
-        "@com_github_aws_aws_sdk_go_v2//aws",
-        "@io_k8s_sigs_controller_runtime//pkg/client",
     ],
 )
