+++ registry/backend/BUILD.bazel
@@ -1,0 +1,16 @@
+load("@rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "backend",
+    srcs = ["backend.go"],
+    importpath = "github.com/bpalermo/aether/registry/backend",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//registry",
+        "//registry/internal/ddb",
+        "//registry/internal/etcd",
+        "//registry/internal/k8s",
+        "@com_github_aws_aws_sdk_go_v2_config//:config",
+        "@io_k8s_sigs_controller_runtime//pkg/client",
+    ],
+)
