+++ registry/registrarclient/BUILD.bazel
@@ -1,0 +1,9 @@
+load("@rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "registrarclient",
+    srcs = ["registrarclient.go"],
+    importpath = "github.com/bpalermo/aether/registry/registrarclient",
+    visibility = ["//visibility:public"],
+    deps = ["//registry/internal/registrar"],
+)
